chore(checkable-input): remove legacy component and update imports
